a whole load of refactoring: WHARF-14 - Added Ivy and Ibiblio resolver - Added supports file URL in wharf resolvers - Added retrieval of sha1 checksum from first head request

+++ wharf-core/src/main/java/org/jfrog/wharf/ivy/resolver/WharfResourceDownloader.java
@@ -22,6 +22,7 @@ import org.apache.ivy.core.module.descriptor.Artifact;
 import org.apache.ivy.plugins.repository.Resource;
 import org.apache.ivy.plugins.repository.ResourceDownloader;
 import org.apache.ivy.plugins.repository.url.URLRepository;
+import org.jfrog.wharf.ivy.resource.WharfUrlResource;
 
 import java.io.File;
 import java.io.IOException;
@@ -33,20 +34,25 @@ public class WharfResourceDownloader implements ResourceDownloader {
 
     private URLRepository extartifactrep = new URLRepository(); // used only to download
 
-    private final IvyWharfResolver resolver;
+    private final WharfResolver resolver;
 
-    public WharfResourceDownloader(IvyWharfResolver resolver) {
+    public WharfResourceDownloader(WharfResolver resolver) {
         this.resolver = resolver;
     }
 
     @Override
     public void download(Artifact artifact, Resource resource, File dest) throws IOException {
+        if (!(resource instanceof WharfUrlResource)) {
+            throw new IllegalArgumentException("The Wharf Resolver manage only WharfUrlResource");
+        }
+        //TODO: [by tc] Don't delete before download, if this part file already exists, manage with a lock
         if (dest.exists()) {
             dest.delete();
         }
         File part = new File(dest.getAbsolutePath() + ".part");
         if (resource.getName().equals(
                 String.valueOf(artifact.getUrl()))) {
+            //TODO: [by tc] in a wharf env this should not happen => throw exception
             if (part.getParentFile() != null) {
                 part.getParentFile().mkdirs();
             }
